Simplification of kernel queue start/stop logic.
We had a local flag that was being used inconsistently to mirror the queue state, but was really brain dead.  sc_devstopped could be off even when the queue was stopped and no matter how many buffers were freed, we would never restart the queue.  This could lead to liveness issues (mostly after a buffer leak caused excessive buffers to be used).  
The kernel has an easy call to find out if the queue is stopped or not, so this checkin uses that.  

We will re-awaken the queue if:

1) we have some buffers we are willing to use for data
2) the channel is available (as opposed to being in DFS CAC)
3) the queue is stopped

This is what we were originally going for with reap counters and sc_dev_stopped and all this other nonsense.  The new logic is much simpler and cleaner.

This also fixes a performance problem where the queue was being re-awakened when no buffers were available resulting in a constant ping-pong of buffers between the kernel and madwifi and a very very heavy CPU utilization at exactly the wrong time.


git-svn-id: http://madwifi-project.org/svn/madwifi/trunk@3503 0192ed92-7a03-0410-a25b-9323aeb14dbd