Eliminate ath_buf's bf_node pointer.
This pointer is entirely redundant with the pointer already in the SKB.  This eliminates an unnecessary source of possible node reference leaks.  In all cases this variable was being populated from the SKB's node pointer and was never referenced outside of the context of processing an skb, for obvious reasons.

Use BF_NI(bf) or SKB_NI(skb) macros to obtain the node of a buffer or skb.


git-svn-id: http://madwifi-project.org/svn/madwifi/trunk@3504 0192ed92-7a03-0410-a25b-9323aeb14dbd