Protect against probing AR2425 if HAL doesn't support it
AR2424 and AR2425 share the same PCI ID.  Probing AR2425 with HAL that
doesn't support it is problematic.  It causes an unhandled NMI and puts
the device to a non-functional state.  After that, even HAL that
supports AR2425 won't be able to use the device until the system is
rebooted.

Reject AR2425 devices if HAL doesn't positively indicate support for
them.


git-svn-id: http://madwifi-project.org/svn/madwifi/trunk@3825 0192ed92-7a03-0410-a25b-9323aeb14dbd