Rewrite ieee80211_input_all()
The logic to prevent calling ieee80211_input() on inactive VAPs was
flawed.  The first VAP was exempt from the check for no reason.

Rewrite to avoid having nested loops.  Give the original skb to the
first VAP, not to the last one, as the first VAP is easier to find in a
loop.  Process the first VAP after others.

Only collect the result from the first VAP.  ieee80211_input() should
return the same value (frame type) for all VAPs except if we call it
incorrectly, e.g. on an inactive VAP.


git-svn-id: http://madwifi-project.org/svn/madwifi/trunk@4081 0192ed92-7a03-0410-a25b-9323aeb14dbd